+++ src/main/java/com/michelin/cio/hudson/plugins/rolestrategy/RoleBasedAuthorizationStrategy.java
@@ -87,7 +87,6 @@ import org.acegisecurity.acls.sid.PrincipalSid;
 import org.apache.commons.lang3.StringUtils;
 import org.jenkinsci.plugins.rolestrategy.AmbiguousSidsAdminMonitor;
 import org.jenkinsci.plugins.rolestrategy.permissions.PermissionHelper;
-import org.jenkinsci.plugins.rolestrategy.permissions.RoleStrategyPermissions;
 import org.kohsuke.accmod.Restricted;
 import org.kohsuke.accmod.restrictions.DoNotUse;
 import org.kohsuke.accmod.restrictions.NoExternalUse;
@@ -1528,7 +1527,7 @@ public class RoleBasedAuthorizationStrategy extends AuthorizationStrategy {
           filterGroups.remove(PermissionGroup.get(Computer.class));
 
           // RoleStrategy permissions
-          filterGroups.remove(PermissionGroup.get(RoleStrategyPermissions.class));
+          filterGroups.remove(PermissionGroup.get(RoleStrategyConfig.class));
           break;
         case SLAVE:
           filterGroups.remove(PermissionGroup.get(Permission.class));
@@ -1536,7 +1535,7 @@ public class RoleBasedAuthorizationStrategy extends AuthorizationStrategy {
           filterGroups.remove(PermissionGroup.get(View.class));
 
           // RoleStrategy permissions
-          filterGroups.remove(PermissionGroup.get(RoleStrategyPermissions.class));
+          filterGroups.remove(PermissionGroup.get(RoleStrategyConfig.class));
 
           // Project, SCM and Run permissions
           filterGroups.remove(PermissionGroup.get(Item.class));
