+++ src/main/resources/com/michelin/cio/hudson/plugins/rolestrategy/RoleStrategyConfig/assign-project-roles.jelly
@@ -29,9 +29,9 @@
   <j:set var="tableid" value="projectRoles"/>
   <j:set var="projectSIDs" value="${it.strategy.getSidEntries(it.strategy.PROJECT)}"/>
   <j:set var="itemGrantedRoles" value="${it.strategy.getGrantedRolesEntries(it.strategy.PROJECT)}"/>
+  <j:set var="neededPermission" value="${it.ITEM_ROLES_ADMIN}"/>
 
-
-  <local:table id="${tableid}" roles="${itemGrantedRoles}" showPattern="true" template="newItemRowTemplate" highlighter="itemTableHighlighter">
+  <local:table id="${tableid}" roles="${itemGrantedRoles}" showPattern="true" template="newItemRowTemplate" highlighter="itemTableHighlighter" neededPermission="${neededPermission}">
     <j:if test="${rspMode == 'rsp-filters'}">
       <div id="itemUserInputFilter" data-table-id="${tableid}" data-initial-size="${projectSIDs.size()}" class="user-filter">
         <f:entry title="${%Filter by User/Group}">
@@ -52,9 +52,9 @@
         </tr>
       </template>
 
-      <l:isAdmin>
+      <l:hasPermission permission="${it.ITEM_ROLES_ADMIN}">
         <local:addButtons sids="${projectSIDs}" tableid="${tableid}" id="newItemRowTemplate" roles="${itemGrantedRoles}" highlighter="itemTableHighlighter"/>
-      </l:isAdmin>
+      </l:hasPermission>
     </j:if>
   </local:table>
