+++ src/main/resources/com/michelin/cio/hudson/plugins/rolestrategy/RoleStrategyConfig/assign-roles.jelly
@@ -28,14 +28,30 @@
 <?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" 
           xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:local="local">
-    <l:layout permission="${app.SYSTEM_READ}" title="${%Assign Roles}" norefresh="true">
+    <l:layout permissions="${it.SYSTEM_READ_AND_SOME_ROLES_ADMIN}" title="${%Assign Roles}" norefresh="true">
       <l:breadcrumb title="Assign Roles"/>
         <j:choose>
-          <j:when test="${app.hasPermission(app.ADMINISTER)}">
-            <j:set var="readOnlyMode" value="false" />
+          <j:when test="${app.hasPermission(it.GLOBAL_ROLES_ADMIN)}">
+            <j:set var="readOnlyModeGlobalRoles" value="false" />
           </j:when>
           <j:otherwise>
-            <j:set var="readOnlyMode" value="true" />
+            <j:set var="readOnlyModeGlobalRoles" value="true" />
+          </j:otherwise>
+        </j:choose>
+        <j:choose>
+          <j:when test="${app.hasPermission(it.ITEM_ROLES_ADMIN)}">
+            <j:set var="readOnlyModeItemRoles" value="false" />
+          </j:when>
+          <j:otherwise>
+            <j:set var="readOnlyModeItemRoles" value="true" />
+          </j:otherwise>
+        </j:choose>
+        <j:choose>
+          <j:when test="${app.hasPermission(it.AGENT_ROLES_ADMIN)}">
+            <j:set var="readOnlyModeAgentRoles" value="false" />
+          </j:when>
+          <j:otherwise>
+            <j:set var="readOnlyModeAgentRoles" value="true" />
           </j:otherwise>
         </j:choose>
         <st:include page="sidepanel.jelly"/>
@@ -43,7 +59,6 @@
           <link rel="stylesheet" href="${rootURL}/plugin/role-strategy/css/role-strategy.css" type="text/css" />
           <script type="text/javascript" src="${rootURL}/plugin/role-strategy/js/table.js" />
           <script type="text/javascript" src="${rootURL}/plugin/role-strategy/js/tableAssign.js" />
-        
           <j:if test="${empty(descriptorPath)}">
             <j:set var="descriptorPath" value="${rootURL}/descriptor/${it.strategy.descriptor.clazz.name}"/>
           </j:if>
@@ -57,13 +72,13 @@
                        data-template-id="${attrs.template}">
 
                   <!-- The first row will show grouping -->
-                  <local:thead roles="${attrs.roles}" showPattern="true"/>
+                  <local:thead roles="${attrs.roles}" showPattern="true" neededPermission="${attrs.neededPermission}"/>
                   <tbody>
                     <tr>
                       <td colspan="100%"><l:spinner text="${%Loading Data}"/></td>
                     </tr>
                   </tbody>
-                  <local:tfoot roles="${attrs.roles}" showPattern="${attrs.showPattern}"/>
+                  <local:tfoot roles="${attrs.roles}" showPattern="${attrs.showPattern}" neededPermission="${attrs.neededPermission}"/>
                 </table>
                 <j:set var="rspMode" value="rsp-navigation"/>
                 <d:invokeBody/>
@@ -72,7 +87,7 @@
             <d:tag name="thead">
               <thead>
                 <tr class="group-row highlight-row">
-                  <l:isAdmin><td class="start" /></l:isAdmin>
+                  <l:hasPermission permission="${attrs.neededPermission}"><td class="start" /></l:hasPermission>
                   <th class="rsp-table__header-column">
                     ${%User/Group}
                   </th>
@@ -90,14 +105,14 @@
                       <span data-html-tooltip="${permissionList}">${role.key.name}</span>
                     </th>
                   </j:forEach>
-                  <l:isAdmin><td class="stop" /></l:isAdmin>
+                  <l:hasPermission permission="${attrs.neededPermission}"><td class="stop" /></l:hasPermission>
                 </tr>
               </thead>
             </d:tag>
             <d:tag name="tfoot">
               <tfoot class="jenkins-hidden">
                 <tr class="group-row highlight-row">
-                  <l:isAdmin><td class="start" /></l:isAdmin>
+                  <l:hasPermission permission="${attrs.neededPermission}"><td class="start" /></l:hasPermission>
                   <th class="rsp-table__footer-column">
                     ${%User/Group}
                   </th>
@@ -115,7 +130,7 @@
                       <span data-html-tooltip="${permissionList}">${role.key.name}</span>
                     </th>
                   </j:forEach>
-                  <l:isAdmin><td class="stop" /></l:isAdmin>
+                  <l:hasPermission permission="${attrs.neededPermission}"><td class="stop" /></l:hasPermission>
                 </tr>
               </tfoot>
             </d:tag>
@@ -156,14 +171,14 @@
             </d:tag>
             <d:tag name="userRow">
               <j:set var="permissionEntry" value="${it.strategy.descriptor.entryFor(attrs.permissionType, attrs.sid)}"/>
-              <l:isAdmin>
+              <l:hasPermission permission="${attrs.neededPermission}">
                 <td class="start">
                     <div class="rsp-remove">
                       <l:icon src="symbol-trash-outline plugin-ionicons-api" class="icon-sm icon-red rsp-table__icon"/>
                     </div>
                 </td>
-              </l:isAdmin>
-	            <td class="left-most">${title}</td>
+              </l:hasPermission>
+              <td class="left-most">${title}</td>
               <j:set var="pattern" value="&lt;b&gt;Pattern&lt;/b&gt; : ${h.escape(r.key.pattern.toString())}"/>
               <j:forEach var="r" items="${it.strategy.getGrantedRolesEntries(attrs.type)}">
                 <j:set var="pattern" value="&lt;b&gt;Pattern&lt;/b&gt; : ${h.escape(r.key.pattern.toString())}"/>
@@ -174,7 +189,7 @@
                   <f:checkbox name="[${r.key.name}]" checked="${r.value.contains(permissionEntry)}" class="rsp-checkbox"/>
                 </td>
               </j:forEach>
-              <l:isAdmin>
+              <l:hasPermission permission="${attrs.neededPermission}">
                 <td class="stop">
                   <div class="rsp-table__cell">
                     <div class="rsp-remove">
@@ -189,7 +204,7 @@
                     </button>
                   </div>
                 </td>
-              </l:isAdmin>
+              </l:hasPermission>
             </d:tag>
           </d:taglib>
 
@@ -229,12 +244,12 @@
             <input type="hidden" name="rolesMapping"/>
             <input type="hidden" name="submit" value="true"/>
           </f:form>
-          <l:isAdmin>
+          <j:if test="${app.hasAnyPermission(it.ADMINISTER_AND_SOME_ROLES_ADMIN)}">
             <f:bottomButtonBar>
               <button id="rsp-roles-save" type="button" class="jenkins-button jenkins-button--primary">${%Save}</button>
               <button id="rsp-roles-apply" type="button" class="jenkins-button" data-message="${%Saved}">${%Apply}</button>
             </f:bottomButtonBar>
-          </l:isAdmin>
+          </j:if>
           <st:adjunct includes="lib.form.confirm" />
         </l:main-panel>
     </l:layout>
